requirements.txt
attrs==19.3.0
Click==7.0
flasgger==0.9.3
Flask==1.1.1
gunicorn==19.9.0
IMDbPY==6.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
lxml==4.4.1
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
pyrsistent==0.15.4
PyYAML==5.1.2
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
zipp==0.6.0
